WA-357: Refactored tests
