+++ app/src/main/java/pm/gnosis/heimdall/di/modules/ApplicationBindingsModule.kt
@@ -11,6 +11,8 @@ import pm.gnosis.heimdall.reporting.CrashTracker
 import pm.gnosis.heimdall.reporting.EventTracker
 import pm.gnosis.heimdall.reporting.impl.FabricCrashTracker
 import pm.gnosis.heimdall.reporting.impl.FabricEventTracker
+import pm.gnosis.heimdall.ui.transactions.view.helpers.DefaultSubmitTransactionHelper
+import pm.gnosis.heimdall.ui.transactions.view.helpers.SubmitTransactionHelper
 import javax.inject.Singleton
 
 @Module
@@ -32,9 +34,12 @@ abstract class ApplicationBindingsModule {
     @Binds
     abstract fun bindsAddressStore(helper: SimpleAddressStore): AddressStore
 
-    // This is unscoped so it will get recreated each time it is injected
     @Binds
     abstract fun bindsSignatureStore(helper: SimpleSignatureStore): SignatureStore
+
+    // This is unscoped so it will get recreated each time it is injected
+    @Binds
+    abstract fun bindsSubmitTransactionHelper(helper: DefaultSubmitTransactionHelper): SubmitTransactionHelper
 
     @Binds
     @Singleton
