+++ app/src/main/java/pm/gnosis/heimdall/ui/transactions/view/review/ReviewTransactionViewModel.kt
@@ -11,6 +11,9 @@ import pm.gnosis.heimdall.data.repositories.models.SafeTransaction
 import pm.gnosis.heimdall.helpers.AddressHelper
 import pm.gnosis.heimdall.helpers.SignatureStore
 import pm.gnosis.heimdall.ui.transactions.view.TransactionInfoViewHolder
+import pm.gnosis.heimdall.ui.transactions.view.helpers.SubmitTransactionHelper
+import pm.gnosis.heimdall.ui.transactions.view.helpers.SubmitTransactionHelper.Events
+import pm.gnosis.heimdall.ui.transactions.view.helpers.SubmitTransactionHelper.ViewUpdate
 import pm.gnosis.heimdall.ui.transactions.view.viewholders.AssetTransferViewHolder
 import pm.gnosis.heimdall.ui.transactions.view.viewholders.GenericTransactionViewHolder
 import pm.gnosis.heimdall.utils.emitAndNext
@@ -28,11 +31,8 @@ import java.math.BigInteger
 import javax.inject.Inject
 
 class ReviewTransactionViewModel @Inject constructor(
-    private val addressHelper: AddressHelper,
-    private val executionRepository: TransactionExecutionRepository,
-    private val signaturePushRepository: PushServiceRepository,
-    private val signatureStore: SignatureStore,
-    private val tokenRepository: TokenRepository
+    private val submitTransactionHelper: SubmitTransactionHelper,
+    private val executionRepository: TransactionExecutionRepository
 ) : ReviewTransactionContract() {
 
     private lateinit var safe: Solidity.Address
@@ -44,6 +44,7 @@ class ReviewTransactionViewModel @Inject constructor(
             cachedState.clear()
         }
         this.safe = safe
+        submitTransactionHelper.setup(safe, ::txParams)
     }
 
     private fun txParams(transaction: SafeTransaction): Single<TransactionExecutionRepository.ExecuteInformation> {
@@ -53,149 +54,5 @@ class ReviewTransactionViewModel @Inject constructor(
     }
 
     override fun observe(events: Events, transactionData: TransactionData): Observable<Result<ViewUpdate>> =
-        transactionViewHolder(transactionData)
-            .emitAndNext(
-                emit = {
-                    DataResult(
-                        ViewUpdate.TransactionInfo(
-                            it
-                        )
-                    )
-                },
-                next = { estimation(events, it) })
-
-    private fun estimation(events: Events, viewHolder: TransactionInfoViewHolder) =
-        viewHolder.loadTransaction().flatMapObservable { tx ->
-            events.retry
-                .subscribeOn(AndroidSchedulers.mainThread())
-                .startWith(Unit)
-                .switchMapSingle { txParams(tx).mapToResult() }
-        }.emitAndNext(
-            emit = {
-                it.map {
-                    ViewUpdate.Estimate(
-                        Wei(
-                            (it.txGas + it.dataGas + BigInteger.valueOf(32000)) * it.gasPrice
-                        ), it.balance
-                    )
-                }
-            },
-            next = { confirmations(events, it) })
-
-    private fun confirmations(events: Events, params: Result<TransactionExecutionRepository.ExecuteInformation>) =
-        (params as? DataResult)?.data?.let {
-            // Once we have the execution information we can setup everything related to requesting and receiving confirmation
-            Observable.merge(
-                requestConfirmation(events, it),
-                observeConfirmationStore(events, it),
-                observeIncomingConfirmations(it)
-            )
-        } ?: Observable.just<Result<ViewUpdate>>(
-            DataResult(ViewUpdate.EstimateError)
-        )
-
-    private fun requestConfirmation(events: Events, params: TransactionExecutionRepository.ExecuteInformation) =
-        if ((params.requiredConfirmation == 1 && params.isOwner))
-            Observable.empty<Result<ViewUpdate>>()
-        else
-            events.requestConfirmations
-                .subscribeOn(AndroidSchedulers.mainThread())
-                .startWith(Unit)
-                .switchMapSingle {
-                    val targets = params.owners - params.sender // TODO: exclude owners that already confirmed
-                    executionRepository.calculateHash(safe, params.transaction, params.txGas, params.dataGas, params.gasPrice)
-                        .flatMapCompletable {
-                            signaturePushRepository.requestConfirmations(
-                                it.toHexString().addHexPrefix(),
-                                safe,
-                                params.transaction,
-                                params.txGas,
-                                params.dataGas,
-                                params.gasPrice,
-                                targets.toSet()
-                            )
-
-                        }
-                        .mapToResult()
-                }
-                .flatMap {
-                    when (it) {
-                        is DataResult -> Observable.just(DataResult(ViewUpdate.ConfirmationsRequested))
-                        is ErrorResult -> Observable.fromArray(DataResult(ViewUpdate.ConfirmationsError), ErrorResult<ViewUpdate>(it.error))
-                    }
-                }
-
-    private fun observeConfirmationStore(events: Events, params: TransactionExecutionRepository.ExecuteInformation) =
-        signatureStore.flatMapInfo(
-            safe, params
-        ).publish {
-            Observable.merge(
-                it.map {
-                    val threshold = params.requiredConfirmation - (if (params.isOwner) 1 else 0)
-                    DataResult(
-                        ViewUpdate.Confirmations(
-                            it.size >= threshold
-                        )
-                    )
-                },
-                it.switchMap { signatures ->
-                    events.submit
-                        .subscribeOn(AndroidSchedulers.mainThread())
-                        .switchMap { submitTransaction(params, signatures) }
-                }
-            )
-        }
-
-    private fun submitTransaction(params: TransactionExecutionRepository.ExecuteInformation, signatures: Map<Solidity.Address, Signature>) =
-        executionRepository.submit(safe, params.transaction, signatures, params.isOwner, params.txGas, params.dataGas, params.gasPrice)
-            .andThen(
-                Observable.just<ViewUpdate>(
-                    ViewUpdate.TransactionSubmitted(
-                        true
-                    )
-                )
-            )
-            .onErrorResumeNext { t: Throwable ->
-                // Propagate error to display snackbar then propagate status
-                Observable.just<ViewUpdate>(
-                    ViewUpdate.TransactionSubmitted(
-                        false
-                    )
-                ).concatWith(Observable.error(t))
-            }
-            .mapToResult()
-
-    private fun observeIncomingConfirmations(params: TransactionExecutionRepository.ExecuteInformation) =
-        signaturePushRepository.observe(params.transactionHash)
-            .flatMap {
-                when (it) {
-                    is PushServiceRepository.TransactionResponse.Confirmed ->
-                        executionRepository.checkConfirmation(safe, params.transaction, params.txGas, params.dataGas, params.gasPrice, it.signature)
-                            .map(signatureStore::add)
-                            .flatMapObservable { Observable.empty<Result<ViewUpdate>>() }
-                            .onErrorResumeNext { e: Throwable -> Observable.just(ErrorResult(e)) }
-                    is PushServiceRepository.TransactionResponse.Rejected ->
-                        executionRepository.checkRejection(safe, params.transaction, params.txGas, params.dataGas, params.gasPrice, it.signature)
-                            .filter { (sender) -> params.owners.contains(sender) }
-                            .map<Result<ViewUpdate>> { DataResult(ViewUpdate.TransactionRejected) }
-                            .toObservable()
-                            .onErrorResumeNext { e: Throwable -> Observable.just(ErrorResult(e)) }
-                }
-            }
-
-    private fun transactionViewHolder(transactionData: TransactionData): Single<TransactionInfoViewHolder> =
-        Single.fromCallable {
-            when (transactionData) {
-                is TransactionData.Generic ->
-                    GenericTransactionViewHolder(safe, transactionData, addressHelper)
-                is TransactionData.AssetTransfer ->
-                    AssetTransferViewHolder(
-                        safe,
-                        transactionData,
-                        addressHelper,
-                        tokenRepository
-                    )
-            }
-        }
-
+        submitTransactionHelper.observe(events, transactionData)
 }
